// explore/explore-ui/projects/explore-ui-core/src/favorite/enums/favorite-type.enum.ts
/**
 * Enum for Favorite types
 */
export enum FavoriteType {
    LAYOUT = 'LAYOUT',
    WORKSPACE = 'WORKSPACE',
    FACTOR_BREAKDOWN = 'FAC_BKD',
    BREAKDOWN = 'BREAKDOWN',
    SECT_CONSTR_BKD = 'SECT_CONSTR_BKD',
    PERFORMANCE_BREAKDOWN = 'PERF_BKD',
    CUSTOM_SEC = 'CUSTOM_SEC',
    REPORT = 'REPORT',
    COLUMN = 'COLUMN',
    SINGLE_REPORT = 'SINGLE_REPORT',
    MULTI_REPORT = 'MULTI_REPORT',
    RISK_REPORT = 'RISK_REPORT',
    COMMITMENT_RISK_REPORT = 'COMMITMENT_RISK_REPORT',
    RETURN_REPORT = 'RETURN_REPORT',
    COMP_RULES = 'COMP_RULES',
    MANDATE_MAP = 'MANDATE_MAP',
    CHART_REPORT = 'CHART_REPORT',
    EXPOST_REPORT = 'EXPOST_REPORT',
    CURATED_REPORTS = 'CURATED_REPORTS',
    ATTRIBUTION_TYPE = 'ATTRIBUTION_TYPE',
    MANDATE = 'MANDATE',
    SCHEDULED_BATCH = 'SCHEDULED_BATCH',
    BATCH_REPORT = 'BATCH_REPORT',
    WHATIF_RULES = 'WHATIF_RULES',
    WHATIF_POS = 'WHATIF_POS',
    ADHOC_PORT = 'ADHOC_PORT',
    ADHOC_PG = 'ADHOC_PG',
    adhocPortfolio = 'adhocPortfolio',
    OPTO_SETTINGS = 'OPTO_SETTINGS'
}